Add common fixups and relocations
%lo(), %hi(), and %pcrel_hi() are supported and test cases have been added to
ensure the appropriate fixups and relocations are generated. I've added an
instruction format field which is used in RISCVMCCodeEmitter to, for instance,
tell whether it should emit a lo12_i fixup or a lo12_s fixup (RISC-V has
two 12-bit immediate encodings depending on the instruction type).

Additionally, fixups and relocations in jal and branch instructions are
supported. Errors are reported in the case of out-of-range or insufficiently
aligned fixups.

Differential Revision: https://reviews.llvm.org/D23568

add checking for out-of-range and insufficiently aligned fixups