Initial codegen support for ALU operations
This roughly constitutes the minimum necessary to get code generation for
simple ALU operations to work. Frame handling, and generation of full function
prologue/epilogues remains unimplemented until a future patch.

Note that RISCVCallingConv.td does not implement the full details of the
RISC-V calling convention at this stage. We also choose not to mark the stack
pointer as a callee-saved register, as the function prologue and epilogue
inserter will manage it instead.

A default SelectionDAGTargetInfo is needed in RISCVSubtarget to support
target-specific generation of libc functions like strcmp.